move enum out to its own file
