+++ packages/arb-token-bridge-ui/src/components/TopNavBar.tsx
@@ -5,11 +5,6 @@ import { twMerge } from 'tailwind-merge'
 import Image from 'next/image'
 import { useTransactionReminderInfo } from './TransactionHistory/useTransactionReminderInfo'
 
-export enum TabParamEnum {
-  BRIDGE = 0,
-  TX_HISTORY = 1
-}
-
 function StyledTab({ children, ...props }: PropsWithChildren) {
   return (
     <Tab
