+++ packages/arb-token-bridge-ui/src/hooks/enum.ts
@@ -1,0 +1,8 @@
+export enum TabParamEnum {
+  BRIDGE = 0,
+  TX_HISTORY = 1
+}
+
+export enum AmountQueryParamEnum {
+  MAX = 'max'
+}
